pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "equity-signals"
version = "0.1.0"
description = "Equity signal generation pipeline using Alpaca and FMP data sources."
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "alpaca-py>=0.13.0",
    "pyarrow>=12.0",
    "pyyaml>=6.0",
    "yfinance>=0.2.40",
]

[project.optional-dependencies]
api = [
    "fastapi[standard]>=0.111.0",
    "gunicorn>=22.0.0",
    "anthropic>=0.28.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "httpx>=0.27.0",        # FastAPI TestClient
]

[project.scripts]
equity-signals        = "equity_signals.cli:validate_config"
equity-universe-scan  = "equity_signals.scripts.run_universe_scan:main"
equity-signal-scan    = "equity_signals.scripts.run_signal_scan:main"
run-weekly-scan       = "scripts.run_weekly_scan:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["equity_signals*", "scripts*"]
